Move CLI fs artifact helpers from shared to modules
Relocate the writeFile/writeJson artifact helpers from src/cli/src/shared/ into src/cli/src/modules/ so the shared bundle only exposes cross-cutting utilities while keeping the legacy shared entry point as a shim.
